增加dice loss

+++ utils/dataset.py
@@ -28,7 +28,7 @@ class BasicDataset(Dataset):
         assert newW > 0 and newH > 0, 'Scale is too small'
         pil_img = pil_img.resize((newW, newH))
 
-        img_nd = np.array(pil_img)
+        img_nd = np.array(pil_img).astype('float32')
 
         if len(img_nd.shape) == 2:
             img_nd = np.expand_dims(img_nd, axis=2)
